+++ base_decorator.rb
@@ -12,3 +12,15 @@ class BaseDecorator < Nameable
     @nameable.correct_name
   end
 end
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.upcase
+  end
+end
+
+class TrimmerDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name[0..9] if @nameable.correct_name.length >= 10
+  end
+end
