--- capitalize_decorator.rb
@@ -1,7 +1,0 @@
-require './base_decorator'
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.upcase
-  end
-end
