--- trimmer_decorator.rb
@@ -1,7 +1,0 @@
-require './base_decorator.rb'
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name[0..9] if @nameable.correct_name.length >= 10
-  end
-end
